Rework the object model
Having all the real objects (Tree, Commit, etc.) contain a RawObject was
quick to get up and running, but then I wanted to start writing objects,
and it was really getting in the way. That is: if I wanted to construct
a commit in memory, I couldn't easily do that, because I didn't have a
raw object to stick inside it.

This commit gets rid of RawObject entirely in favor of the Object enum
(which now contain structs), and reworks the GitObject trait to work
with that. I think this is going to be much easier to work with, in the
long run.

(This compiles, but does not yet fully work; some things are marked as
todo.)