Pass commands a new Context struct
This gives them a handle to stdout, and also to the repository.
`util::find_repo` worked fine, but won't work well for testing, because
it depends on an env var or the current working directory, neither of
which work well in threaded environments.

As of this change, `find_repo` now returns an Option (rather than a
Result), and `ctx.repo()` promotes it to a result. That's because not
all commands _need_ the repository (notably: init!), and so sometimes
passing in None to the dispatcher is sensible.

This also gets rid of the recently introduced Stdout struct; the
context will handle printing too. Plus, if later I decide I
need something else, I can just stick it in the context!