Move a bunch of commit logic into repo
This is another mishmash commit. I want to be able to easily test
repository actions, and running commit is kind of a pain, because (right
now) it depends on the env for author name/email settings. This moves
most of that logic into the repository impl, and also reworks it a bit.
(Person now contains a date, for example).

There's also some further stuff here about canonicalization/
relativization of paths, which is tedious but now I think mostly
correct.