status: do not hold a reference to the index
This was making the implementation of diff pretty annoying. Before, the
Status struct held a mutable reference to the index (because it writes
stat information to it). In the diff command, later we'd want to read
the index, which we could not do, because there was already a mutable
reference to it.

This is silly, but: now does the same thing it did before, but the
struct it returns is just the diff data, with no references.