Rewrite the command abstraction (again)
That is: this is no longer a trait which required a bunch of empty
structs. I think this will be more flexible, because if you want to make
some kind of helper struct that only lives for the lifetime of run(),
you can do so without faffing with a bunch of lifetimes.