Myers diff: consume self to avoid string cloning
Myers::new() takes two strings, and moves them into a WrappingVec. This
commit addes WrappingVec::take(), which moves those strings out of the
wrapping vec (and replaces them with a default, which we can do because
we're guaranteed to have one). This means that when we generate the
diff, we can move those strings out of ourselves and into the vec of
Edits, which is both faster and saves memory.